requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-ollama
python-dotenv
pypdf